// src/client/models/vehicle-journey.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Navitia API - SNCF
 * Complete API documentation for Navitia/SNCF public transportation API. Based on https://doc.navitia.io
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Disruption } from './disruption';
// May contain unused imports in some cases
// @ts-ignore
import type { JourneyPattern } from './journey-pattern';
// May contain unused imports in some cases
// @ts-ignore
import type { Link } from './link';
// May contain unused imports in some cases
// @ts-ignore
import type { StopTime } from './stop-time';

export interface VehicleJourney {
    /**
     * Vehicle journey ID
     */
    'id'?: string;
    /**
     * Vehicle journey name
     */
    'name'?: string;
    /**
     * Headsign
     */
    'headsign'?: string;
    'journey_pattern'?: JourneyPattern;
    'stop_times'?: Array<StopTime>;
    'disruptions'?: Array<Disruption>;
    'links'?: Array<Link>;
}